requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
feedparser>=6.0.10
python-dotenv>=1.0.0

# Optional: For local LLM generation (if using Hugging Face)
# transformers>=4.35.0
# torch>=2.1.0

# CLI enhancements
rich>=13.7.0
prompt-toolkit>=3.0.0

# Web interface
streamlit>=1.28.0

